Move the output readout to a seperate file
